fix: Refactored request and status generation file
